+++ htdocs/view/delete.php
@@ -1,7 +1,8 @@
 <?php
 /**
  * Mahara: Electronic portfolio, weblog, resume builder and social networking
- * Copyright (C) 2006-2008 Catalyst IT Ltd (http://www.catalyst.net.nz)
+ * Copyright (C) 2006-2009 Catalyst IT Ltd and others; see:
+ *                         http://wiki.mahara.org/Contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -20,7 +21,7 @@
  * @subpackage core
  * @author     Catalyst IT Ltd
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL
- * @copyright  (C) 2006-2008 Catalyst IT Ltd http://catalyst.net.nz
+ * @copyright  (C) 2006-2009 Catalyst IT Ltd http://catalyst.net.nz
  *
  */
 
@@ -71,10 +72,15 @@ $smarty->assign('form', $form);
 $smarty->display('view/delete.tpl');
 
 function deleteview_submit(Pieform $form, $values) {
-    global $SESSION, $viewid, $groupid, $institution;
+    global $SESSION, $USER, $viewid, $groupid, $institution;
     $view = new View($viewid, null);
-    $view->delete();
-    $SESSION->add_ok_msg(get_string('viewdeleted', 'view'));
+    if (View::can_remove_viewtype($view->get('type')) || $USER->get('admin')) {
+        $view->delete();
+        $SESSION->add_ok_msg(get_string('viewdeleted', 'view'));
+    }
+    else {
+        $SESSION->add_error_msg(get_string('cantdeleteview', 'view'));
+    }
     if ($groupid) {
         redirect('/view/groupviews.php?group='.$groupid);
     }
